Move BackButton to navigation and remove unused files
