+++ src/components/scenes/GuiPluginLegacyScene.js
@@ -13,7 +13,6 @@ import { sendConfirmationUpdateTx } from '../../actions/SendConfirmationActions'
 import { selectWallet } from '../../actions/WalletActions'
 import s from '../../locales/strings.js'
 import T from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import { BackButton } from '../../modules/UI/components/Header/Component/BackButton.ui'
 import { PluginBridge, pop as pluginPop } from '../../modules/UI/scenes/Plugins/api'
 import * as UI_SELECTORS from '../../modules/UI/selectors.js'
 import type { GuiMakeSpendInfo } from '../../reducers/scenes/SendConfirmationReducer.js'
@@ -21,6 +20,7 @@ import { THEME } from '../../theme/variables/airbitz.js'
 import { type GuiPlugin } from '../../types/GuiPluginTypes.js'
 import { type GuiWallet } from '../../types/types.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { BackButton } from '../navigation/BackButton.js'
 import { showError } from '../services/AirshipInstance.js'
 
 type Props = {
